config: create Config objects with JK support
This teaches install_config_module (and the underlying code) to
understands Config objects. Additionally we've added a JK option to this
which resolves the JK.

This config gets stored within the _ConfigEntry class and is evaluated
when __getattr__ is called. If justknobs is set, it'll call
justknobs_check to see the result.

Due to preceeding work, basically everything works correctly here and we
had to update a couple of tests, and modify the getattr behaviour.

Note that we are updating the justknob_check function to support a
default option, to make default work.

ghstack-source-id: f3fbc970ca95458bf75e38a5249c1296e1c8b4cc
Pull Request resolved: https://github.com/pytorch/pytorch/pull/138766